requirements.txt
autogen~=0.3
faiss-cpu
sentence-transformers
python-dotenv
fastapi
uvicorn
